move tft control logic to separate script
